feat: refactor error and content component names
