+++ packages/author-profile/__tests__/android/author-profile.test.js
@@ -1,3 +1,3 @@
-import shared from "../author-profile-content.native";
+import shared from "../author-profile-list-content.native";
 
 describe("AuthorProfile tests on android", shared);
