+++ packages/author-profile/src/author-profile.js
@@ -6,8 +6,8 @@ import {
   AuthorArticlesWithImagesProvider
 } from "@times-components/provider";
 import { ratioTextToFloat } from "@times-components/utils";
-import AuthorProfileContent from "./author-profile-content";
-import AuthorProfileError from "./author-profile-list-page-error";
+import AuthorProfileListContent from "./author-profile-list-content";
+import AuthorProfileListPageError from "./author-profile-list-page-error";
 import { propTypes, defaultProps } from "./author-profile-prop-types";
 import authorProfileTrackingContext from "./author-profile-tracking-context";
 
@@ -25,12 +25,12 @@ const AuthorProfile = ({
   slug
 }) => {
   if (error) {
-    return <AuthorProfileError refetch={refetch} />;
+    return <AuthorProfileListPageError refetch={refetch} />;
   }
 
   if (isLoading) {
     return (
-      <AuthorProfileContent
+      <AuthorProfileListContent
         articlesLoading
         imageRatio={ratioTextToFloat("3:2")}
         isLoading
@@ -73,7 +73,7 @@ const AuthorProfile = ({
         refetch: refetchArticles,
         variables: { imageRatio = "3:2" }
       }) => (
-        <AuthorProfileContent
+        <AuthorProfileListContent
           articles={get(data, "articles.list", [])}
           articlesLoading={articlesLoading}
           biography={biography}
